feat: root vocabulary alignment, thinking budget, and test refactoring
Promote cross-cutting option constants (OptionSystemPrompt, OptionMaxTurns)
from claude/ to root package, add OptionThinkingBudget as root-level option,
and align all backends to the "root is language, backends are dialect"
design philosophy documented in DESIGN.md.

Key changes:
- Add OptionSystemPrompt, OptionMaxTurns, OptionThinkingBudget to root
- Implement thinking budget in Claude backend (--max-thinking-tokens)
- Extract validatePositiveInt/appendPositiveInt helpers for option handling
- Add strict ResumeArgs validation for OptionMaxTurns (parity with thinking budget)
- Change IsDelta() from explicit switch to convention-based HasSuffix("_delta")
- Fix MessageThinking/MessageThinkingDelta docstrings per empirical findings
- Split 1703-line claude_test.go into args_test.go, parse_test.go, claude_test.go
- Add DESIGN.md codifying root-is-language design philosophy
- Add stream-inspector example and shared display package
