+++ agentrun.go
@@ -11,6 +11,18 @@
 //   - [Message] — structured output from agent processes
 //   - [Option] — functional options for [Engine.Start]
 //
+// # Vocabulary
+//
+// The root package defines the shared vocabulary for all backends:
+//
+//   - Output vocabulary: [MessageType] constants define what agents produce
+//   - Input vocabulary: Option* constants define cross-cutting session configuration
+//
+// Backends translate this vocabulary into their wire format. Cross-cutting
+// concepts (system prompts, turn limits, thinking budgets) are defined here
+// as well-known [Session.Options] keys. Backend-specific concepts remain in
+// their respective packages.
+//
 // # Quick Start
 //
 //	engine := cli.NewEngine(claude.New())
