+++ session.go
@@ -1,5 +1,32 @@
 package agentrun
 
+// Well-known option keys for Session.Options.
+//
+// These keys define cross-cutting concepts that multiple backends may
+// support. Backends silently ignore keys they don't recognize.
+// Backend-specific options are defined in their own packages.
+const (
+	// OptionSystemPrompt sets the system prompt for the session.
+	// Value is the raw prompt string.
+	OptionSystemPrompt = "system_prompt"
+
+	// OptionMaxTurns limits the number of agentic turns per invocation.
+	// Value is a positive integer string (e.g., "5").
+	OptionMaxTurns = "max_turns"
+
+	// OptionThinkingBudget controls the model's thinking/reasoning output.
+	// When set to a positive integer string (e.g., "10000"), backends that
+	// support extended thinking will emit MessageThinking and
+	// MessageThinkingDelta messages with the model's reasoning content.
+	//
+	// The value interpretation is backend-specific:
+	//   - Claude CLI: token count, maps to --max-thinking-tokens
+	//   - Other backends: may accept token counts, levels, or other formats
+	//
+	// When empty or absent, thinking output is disabled (backend default).
+	OptionThinkingBudget = "thinking_budget"
+)
+
 // Session is the minimal session state passed to engines.
 //
 // Session is a value type — it carries identity and configuration but
@@ -21,8 +48,9 @@ type Session struct {
 	// Prompt is the initial prompt or message for the session.
 	Prompt string `json:"prompt,omitempty"`
 
-	// Options holds backend-specific key-value configuration.
-	// CLI backends use this for flags like permission mode.
-	// API backends use this for endpoint configuration.
+	// Options holds session configuration as key-value pairs.
+	// Use well-known Option* constants for cross-cutting features
+	// (e.g., OptionSystemPrompt, OptionMaxTurns). Backend-specific
+	// options are defined in their respective packages.
 	Options map[string]string `json:"options,omitempty"`
 }
